+++ ui/src/shared/components/InputClickToEdit.tsx
@@ -8,6 +8,7 @@ interface Props {
   disabled?: boolean
   tabIndex?: number
   placeholder?: string
+  appearAsNormalInput: boolean
 }
 
 interface State {
@@ -85,9 +86,19 @@ class InputClickToEdit extends PureComponent<Props, State> {
 
   render() {
     const {isEditing} = this.state
-    const {wrapperClass, disabled, tabIndex, placeholder, value} = this.props
+    const {
+      wrapperClass: wrapper,
+      disabled,
+      tabIndex,
+      placeholder,
+      value,
+      appearAsNormalInput,
+    } = this.props
 
-    const divStyle = value ? 'input-cte' : 'input-cte__empty'
+    const wrapperClass = `${wrapper}${appearAsNormalInput
+      ? ' input-cte__normal'
+      : ''}`
+    const defaultStyle = value ? 'input-cte' : 'input-cte__empty'
 
     return disabled
       ? <div className={wrapperClass}>
@@ -107,16 +118,16 @@ class InputClickToEdit extends PureComponent<Props, State> {
                 autoFocus={true}
                 onFocus={this.handleFocus}
                 tabIndex={tabIndex}
-                placeholder={placeholder}
+                spellCheck={false}
               />
             : <div
-                className={divStyle}
+                className={defaultStyle}
                 onClick={this.handleInputClick}
                 onFocus={this.handleInputClick}
                 tabIndex={tabIndex}
               >
                 {value || placeholder}
-                <span className="icon pencil" />
+                {appearAsNormalInput || <span className="icon pencil" />}
               </div>}
         </div>
   }