+++ ui/webpack/dev.config.js
@@ -17,7 +17,7 @@ const babelLoader = {
   loader: 'babel-loader',
   options: {
     cacheDirectory: true,
-    presets: ['env', 'react', 'stage-0'],
+    presets: [['env', {modules: false}], 'react', 'stage-0'],
   },
 }